+++ tests/tests.h
@@ -21,5 +21,6 @@ void suite_linked_list(void);
 void suite_ring_buffer(void);
 void suite_mutex(void);
 void suite_semaphore(void);
+void suite_atomic_queue(void);
 
 #endif /* D4NP_TESTS_H */